Linear Regression on Boston Housing
As a first experience with Linear Regression, it was very understandable
after studying some classification problems. In this particular problem,
our model should estimate the pricing of houses in boston according to
features in the dataset (like crime, age of the owner, etc).

The next step is to use a more flexible method (Gradient Boosting Tree),
to implement the same problem and see what happens to the MSE (current:
15,0110, used on the link below: 28,5414).

This algorithm was based on:
https://medium.com/@haydar_ai/learning-data-science-day-9-linear-regression-on-boston-housing-dataset-cd62a80775ef
and the plot configs were found on:
https://stories.thinkingmachin.es/matplotlib_tutorial/.