Gradient Boosting Regression Trees on Boston Housing
So after using Linear Regression, the results were not optimal, but they
served as a good start.
Reading some articles on Medium about Kaggle, it was clear that the most
dominant method is surely GBM (Gradient Boosting Machines) with
different models, so it's time for me to learn how to tune parameters
and implement a model with a simple dataset (Boston Housing).
Output:
Model Accuracy: 0.9038
MSE: 6.0045
CV Scores: 0.60 (+/- 0.43)