Changed metrics for better assessment
R^2 is the chosen one for this dataset, as it was used on some online courses. But RMSE also give some insights on the standard deviation of what is not represented by the model.

The output of this model is:
RMSE: 3.8744
R^2 score: 0.7545
CV Scores: -0.15 (+/- 1.98)

Which is the basemodel for further improvements.