GradientBoostingRegressor improved by a tutorial
Using this link:
https://www.analyticsvidhya.com/blog/2016/02/complete-guide-parameter-tuning-gradient-boosting-gbm-python/

We're able to improve our model a lot by tuning the GBM parameters. It was useful to create functions to remove clutter and the metrics are up-to-date with last commit.

The output of this model is:
RMSE: 2.1041
R^2 score: 0.9117
CV Scores: 0.61 (+/- 0.52)

This means our model can represent 91% of our data, but there're lots of rooms for improvements.

Next steps:
    - Implement XGBRegressor;
    - Feature Engineering (post-poned);
    - Stacking (need to learn first how to).