Implementation of XGBoost won on BostonHousing (for now)
XGBRegressor far exceeded my epectations on Boston Housing, achieving 93,4% in R^2 score. This is a remarkable improvement, given the CV score also rised in 0.16 (more than 25%).

This algorithm is one of the latest and best for an infinitude of reasons, but for now, let's see what's been done.

Parameter tuning:
 https://www.analyticsvidhya.com/blog/2016/03/complete-guide-parameter-tuning-xgboost-with-codes-python/
    https://cambridgespark.com/content/tutorials/hyperparameter-tuning-in-xgboost/index.html

It's really important to notice no feature engineering was used yet, but in some articles it's recommended not to, as XGBoost takes care of all features proportionally to their importances and also has regularization built-in, so the only problem would be time if the dataset was too big (not the case).

Next steps:
    - Stacking (need to learn how, resources: http://blog.kaggle.com/2016/12/27/a-kagglers-guide-to-model-stacking-in-practice/ , https://burakhimmetoglu.com/2016/12/01/stacking-models-for-improved-predictions/ , https://www.kdnuggets.com/2017/02/stacking-models-imropved-predictions.html )