Implementing RandomForestRegressor on BostonHousing
After finishing XGBRegressor, I've moved to RandomForestRegressor as it's a common implementation on Regression problems.

Kinda used : https://towardsdatascience.com/hyperparameter-tuning-the-random-forest-in-python-using-scikit-learn-28d2aa77dd74

Most of the tuning was based on XGB hyperparameters, so the article was a good guide.

This algorithm is used to reduce variance, but it cannot change bias, It uses fully grown decision trees and various subsamples of the data, using averaging to improve the predictive accuracy and control over-fitting.

The results were almost as good as XGB's. It'll be used as a baseline model to stacking.