Stacking implementation on BostonHousing
As described on the code, stacking was the natural path on our algorithm route, as it combines multiple models outputs as features to increase linear independence of the dataset prediction.

It can be used to make voting ensemble of best models, and I did it with XGB, improving the results of the 2nd level predictions to the highest possible.

For now, the Boston Housing was an excellent toy dataset and I've learned a lot of new techniques, models and python usage. I think I'm ready to try new models and see if what I learned is applicable to other implementations.

The ensemble algorithm used here is them same of Eike Dehling's on https://www.kaggle.com/eikedehling/stack-of-svm-elasticnet-xgboost-rf-0-55/code . I changed variable names for greater readability and added a description for newbs (like me).

Going forward, next problems are going to be :

-
 https://www.kaggle.com/jameslko/gun-violence-data

- https://data.world/databeats/household-power-consumption