Blue book for Bulldozers: pre-processing and EDA
This competition on Kaggle came to my hands by Fast.AI's course on Machine Learning. It's being an excellent experience to learn about EDA and I'll dive soon into Feature Engineering and modeling. The Jupyter Notebook is self-explanatory and will be updated within time. The data can be found on: https://www.kaggle.com/c/bluebook-for-bulldozers